wip : 머지 및 ParticleEmitterInstances 클래스 정리
